# GoMailFlows/docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiTechCloud MailFlows API
  version: 1.0.0
  description: |
    HiTechCloud MailFlows - Enterprise Email SaaS Platform API
    
    ## Features
    - **Domain Management**: Register and manage email domains with automatic DNS provisioning
    - **Email Accounts**: Create and manage email mailboxes
    - **SMTP Configuration**: Per-domain or global SMTP relay configuration
    - **JMAP Support**: Full JMAP protocol (RFC 8620/8621) for modern email access
    - **Email Marketing**: Campaign management, subscriber lists, analytics
    - **SaaS Packages**: Flexible service plans with granular feature control
    - **Admin Panel**: Complete system administration capabilities
    - **DNS Templates**: Customizable DNS record templates for domain provisioning
    - **MCP Integration**: Model Context Protocol support for AI-powered management
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Obtain tokens via `/api/v1/auth/login` or `/api/v1/auth/register`.
    
    ## Rate Limiting
    API endpoints are rate-limited. Limits vary by endpoint and user package.
  contact:
    name: HiTechCloud Support
    url: https://hitechcloud.io/support
    email: support@hitechcloud.io
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.mailflows.hitechcloud.io/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /domains:
    get:
      tags: [domains]
      summary: List domains
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verifying, active, failed, suspended]
      responses:
        '200':
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Domain'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer

    post:
      tags: [domains]
      summary: Register a new domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainRequest'
      responses:
        '201':
          description: Domain created with DNS records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '409':
          description: Domain already registered
  /domains/{id}:
    get:
      tags: [domains]
      summary: Get domain details with DNS records
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Domain details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainWithRecords'

    delete:
      tags: [domains]
      summary: Delete domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Domain deleted
  /domains/{id}/verify:
    post:
      tags: [domains]
      summary: Trigger domain verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainCheckList'
  /domains/{id}/dns/regenerate:
    post:
      tags: [domains]
      summary: Regenerate DNS records
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: New DNS records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DNSRecord'

  # ============================================================
  # Email Accounts
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/register
  schemas:
...
    Domain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [pending, verifying, active, failed, suspended]
        mx_verified:
          type: boolean
        spf_verified:
          type: boolean
        dkim_verified:
          type: boolean
        dmarc_verified:
          type: boolean
        dkim_selector:
          type: string
        max_mailboxes:
          type: integer
        mailbox_count:
          type: integer
        catch_all_address:
          type: string
          nullable: true
        smtp_config_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateDomainRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: FQDN domain name (e.g., example.com)
        template_id:
          type: string
          format: uuid
          description: DNS template ID to use
    DomainWithRecords:
      allOf:
        - $ref: '#/components/schemas/Domain'
        - type: object
          properties:
            dns_records:
              type: array
              items:
                $ref: '#/components/schemas/DNSRecord'
            check_list:
              $ref: '#/components/schemas/DomainCheckList'
    DNSRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [MX, A, AAAA, CNAME, TXT, SRV, CAA, NS]
        host:
          type: string
        value:
          type: string
        priority:
          type: integer
          nullable: true
        ttl:
          type: integer
        is_managed:
          type: boolean
        verified:
          type: boolean
    DomainCheckList:
      type: object
      properties:
        mx_record:
          $ref: '#/components/schemas/CheckItem'
        spf_record:
          $ref: '#/components/schemas/CheckItem'
        dkim_record:
          $ref: '#/components/schemas/CheckItem'
        dmarc_record:
          $ref: '#/components/schemas/CheckItem'
        mta_sts:
          $ref: '#/components/schemas/CheckItem'
        tls_rpt:
          $ref: '#/components/schemas/CheckItem'
    CheckItem:
      type: object
      properties:
        configured:
          type: boolean
        required:
          type: boolean
        record:
          type: string
        value:
          type: string
        message:
          type: string
...